refactor(ffi): extract c_helpers to separate module
Move C Helper functions to dedicated c_helpers.rs module.
