+++ src/ffi/mod.rs
@@ -1,3 +1,5 @@
+pub mod c_helpers;
 pub mod constants;
 
+pub use c_helpers::{cast_string, enabled, present, success, to_c_bool, to_c_result};
 pub(crate) use constants::*;
